:alien: Applied StyleCop principles.

+++ src/maple-fighters/Assets/Maple Fighters/Scripts/Network/Peer Logics/AuthorizationPeerLogic.cs
@@ -7,10 +7,17 @@ namespace Scripts.Services
 {
     public sealed class AuthorizationPeerLogic : PeerLogicBase, IAuthorizationPeerLogicAPI
     {
-        public async Task<AuthorizeResponseParameters> Authorize(IYield yield, AuthorizeRequestParameters parameters)
+        public async Task<AuthorizeResponseParameters> Authorize(
+            IYield yield,
+            AuthorizeRequestParameters parameters)
         {
-            return await ServerPeerHandler.SendOperation<AuthorizeRequestParameters, AuthorizeResponseParameters>
-                (yield, (byte)AuthorizationOperations.Authorize, parameters, MessageSendOptions.DefaultReliable());
+            return 
+                await ServerPeerHandler
+                    .SendOperation<AuthorizeRequestParameters, AuthorizeResponseParameters>(
+                        yield,
+                        AuthorizationOperations.Authorize,
+                        parameters,
+                        MessageSendOptions.DefaultReliable());
         }
     }
 }
